requirements.txt
fastapi
uvicorn
pyzmq
pyjwt
msgpack
protobuf
circuitbreaker
structlog
cryptography
prometheus_client
